+++ CHANGELOG.md
@@ -9,6 +9,9 @@ Unreleased section should follow [Release Toolkit](https://github.com/newrelic/r
 
 ## Unreleased
 
+### bugfix
+- A wrong custom query could cause a panic. Now the error is managed properly.
+ 
 ## v3.8.0 - 2024-06-17
 
 ### 🛡️ Security notices